requirements.txt
fastapi==0.111.0
pydantic==2.7.1
python-dotenv==1.0.1
SQLAlchemy==2.0.29
